Move question out of questions dir
It makes more sense for the JSON to sit alone, as they’re
config files
