+++ server/forms/steps/step.ts
@@ -3,7 +3,7 @@ import { validate, ValidationError } from 'class-validator'
 import type { ReferralApplicationBody, ErrorMessages } from '../interfaces'
 import type { AllowedStepNames, AllowedSectionNames, Dto } from './index'
 import type { ReferralApplication } from '../referral-application.form'
-import Question from '../questions/question'
+import Question from '../question'
 
 export default abstract class Step {
   errors: ErrorMessages
